Massive patch, read on for details
The biggest takeaway here is that you can now create events using
the bot. They're stored in the psql database. The bot is able to
create the embed message, which includes the host's avatar, the
game's cover art, and a local display of the event time.

This patch of course includes the entire skeleton of what it took
to make that happen. This includes a command library, with a
somewhat easy way to create and register new commands. It also
includes some classes related to generating embeds for Discord,
which is currently just for the event, but will be expanded upon
for the help messages, and whatever else might come in the future
for ScrubBot.

Additionally, added a couple utility classes, one for randomly
generating a color that can be used with Discord embeds, and
another for interfacing with the IGDB.