Modified command context, removed event commands, added roll
Changed how context is passed around for a command, given how
useless the previous wrapper was.

Removed the info, start, and players commands for events. They
will be obsolete soon with the reactions update.

Added a new standalone command section, with the first one allowing
for arbitrary rolls, for my RPG leaning folks.

As a note for myself in the future, I've been making a number of
changes in separate efforts, so it's POSSIBLE this is a partially
broken patch. So we're going to leave it in develop until everything
is done.