Added event waitlists
Kind of on a roll tonight, I added the ability to have event
waitlists. If enabled on event creation, any players joining
over the max player count will be added to a waitlist. When a
player leaves, the first player from the waitlist will be
automatically added to the player list.

Updated the event embed to reflect the new waitlist. The embed
will display the first 3 users in the waitlist. Anything beyond
that will be shown as just a number.